requirements_math.txt
# Plataforma de Matemáticas - Backend API
# Requerimientos para la plataforma educativa

fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
pytest==8.3.2
httpx==0.27.2

# Base de datos
alembic==1.13.2

# Validaciones adicionales
email-validator==2.2.0

# Documentación
markdown==3.7
